pyproject.toml
[project]
name = "synthetic_med_models"
packages = [
    { include = "medDataBlender",  from = "." },
]
version = "1.0"
description = "Framework for the creation of synthetic medical data"
authors = [
    {name = "C.Massaroni <cristiano.massaroni@moveax.it>"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3 (>=1.34.0,<2.0.0)",
    "certifi (==2024.8.30)",
    "charset-normalizer (==3.4.0)",
    "contourpy (==1.3.1)",
    "cycler (==0.12.1)",
    "diffusers (==0.31.0)",
    "filelock (==3.13.1)",
    "fonttools (==4.54.1)",
    "fsspec (==2024.2.0)",
    "huggingface-hub (==0.26.5)",
    "idna (==3.10)",
    "imageio (==2.36.0)",
    "importlib-metadata (==8.5.0)",
    "importlib-resources (==6.4.5)",
    "jinja2 (==3.1.3)",
    "joblib (==1.4.2)",
    "kiwisolver (==1.4.7)",
    "lazy-loader (==0.4)",
    "lz4 (==4.4.3)",
    "markupsafe (==2.1.5)",
    "matplotlib (==3.9.2)",
    "mpmath (==1.3.0)",
    "networkx (==3.2.1)",
    "nibabel (==5.3.2)",
    "numpy (<2)",
    "opencv-python (==4.10.0.84)",
    "packaging (==24.2)",
    "pillow (==11.0.0)",
    "pyparsing (==3.2.0)",
    "python-dateutil (==2.9.0.post0)",
    "pyyaml (==6.0.2)",
    "regex (==2024.11.6)",
    "requests (==2.32.3)",
    "safetensors (==0.4.5)",
    "scikit-image (==0.24.0)",
    "scikit-learn (==1.6.1)",
    "scipy (==1.14.1)",
    "six (==1.16.0)",
    "sympy (==1.13.1)",
    "threadpoolctl (==3.5.0)",
    "tifffile (==2024.9.20)",
    "tokenizers (==0.21.0)",
    "tqdm (==4.67.1)",
    "transformers (==4.47.0)",
    "urllib3 (==2.2.3)",
    "zipp (==3.21.0)",
    "typing-extensions (>=4.10.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "wfdb (>=4.2.0,<5.0.0)",
    "einops (>=0.8.1,<0.9.0)",
    "fastdtw (>=0.3.4,<0.4.0)",
    "psycopg2-binary (>=2.9.0,<3.0.0)",
    "torch (==2.0.1)",
    "torchvision (==0.15.2)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"


